+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-# Change Log
+## [x.y.z](https://github.com/sushichop/cordova-plugin-wifi-manager/releases/tag/x.y.z) (yyyy-mm-dd)
+
+- Update `package.json` and plugin interfaces. [#24](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/24) [#24](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/24)
 
 ## [0.3.0](https://github.com/sushichop/cordova-plugin-wifi-manager/tree/0.3.0) (2020-10-25)
 
-- Add AndroidQ compatibility [#14](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/14)
-- Add Cordova Paramedic Tests [#15](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/15) [#16](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/16)
+- Add AndroidQ compatibility. [#14](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/14) [#14](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/14)
+- Add Cordova Paramedic Tests. [#15](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/15) [#16](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/16)
 
 ## [0.2.1](https://github.com/sushichop/cordova-plugin-wifi-manager/tree/0.2.1) (2020-05-25)