+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [x.y.z](https://github.com/sushichop/cordova-plugin-wifi-manager/releases/tag/x.y.z) (yyyy-mm-dd)
+## [0.4.0](https://github.com/sushichop/cordova-plugin-wifi-manager/releases/tag/0.4.0) (2021-07-05)
 
 - Update `package.json` and plugin interfaces. [#24](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/24)
 - Does not add json object in success callback. [#25](https://github.com/sushichop/cordova-plugin-wifi-manager/pull/25)